Add files via upload
Added a "triangle ray-cast" approach, dynamic height search, increased efficiency of the algorithm, and a few more useful changes. The folder does not contain the test iteration results folder, as it contained too many files due to images of inspections to upload. The folder will be automatically re-generated after the algorithm is run.